requirements.txt
surlex
django-tinymce
-e git://github.com/dokterbob/django-metadata.git#egg=django-metadata
-e git://github.com/dokterbob/django-multilingual-model.git#egg=django-multilingual-model
sorl-thumbnail
